// src/shared/ui/Stack/Flex/types.ts
export type TFlexDirections = 'row' | 'column' | 'columnReverse' | 'rowReverse';
export type TFlexWrap = 'wrap' | 'nowrap' | 'wrapReverse';
export type TFlexJustify =
  | 'center'
  | 'start'
  | 'end'
  | 'spaceAround'
  | 'spaceBetween';
export type TFlexAlign = 'center' | 'start' | 'end' | 'stretch' | 'baseline';
export type TFlexGap = '4' | '8' | '12' | '16' | '20' | '24' | '32';